Moved end modal to separated file
